Delete 코드 완료, Rest API 구축 완료

+++ src/main/java/site/metacoding/white/service/BoardService.java
@@ -8,6 +8,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import site.metacoding.white.domain.Board;
 import site.metacoding.white.domain.BoardRepository;
 import site.metacoding.white.domain.UserRepository;
@@ -18,6 +19,7 @@ import site.metacoding.white.dto.BoardRespDto.BoardDetailRespDto;
 import site.metacoding.white.dto.BoardRespDto.BoardSaveRespDto;
 import site.metacoding.white.dto.BoardRespDto.BoardUpdateRespDto;
 
+@Slf4j
 @RequiredArgsConstructor
 @Service
 public class BoardService {
@@ -76,7 +78,12 @@ public class BoardService {
 
     @Transactional
     public void deleteById(Long id) {
-        boardRepository.deleteById(id);
+        Optional<Board> boardOP = boardRepository.findById(id);
+        if (boardOP.isPresent()) {
+            boardRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("해당 " + id + "로 삭제를 할 수 없습니다.");
+        }
     }
 
 }
